Move patterns into Facet.Core.Term.
