+++ src/Facet/Core/Term.hs
@@ -2,6 +2,12 @@ module Facet.Core.Term
 ( -- * Term variables
   Var(..)
 , unVar
+  -- * Patterns
+, ValuePattern(..)
+, Pattern(..)
+, pvar
+, pcon
+, fill
   -- * Term values
 , Value(..)
 , Elim(..)
@@ -25,7 +31,7 @@ import           Data.Foldable (asum, foldl')
 import qualified Data.IntMap as IntMap
 import           Data.Semialign.Exts (zipWithM)
 import           Data.Text (Text)
-import           Facet.Core
+import           Data.Traversable (mapAccumL)
 import qualified Facet.Core.Type as T
 import           Facet.Name
 import           Facet.Stack
@@ -43,6 +49,31 @@ unVar :: (Q Name -> b) -> (a -> b) -> Var a -> b
 unVar f g = \case
   Global  n -> f n
   Free    n -> g n
+
+
+-- Patterns
+
+data ValuePattern a
+  = PWildcard
+  | PVar a
+  | PCon (Q Name :$ ValuePattern a)
+  deriving (Eq, Foldable, Functor, Ord, Show, Traversable)
+
+data Pattern a
+  = PEff (Q Name) (Stack (Pattern a)) a
+  | PAll a
+  | PVal (ValuePattern a)
+  deriving (Eq, Foldable, Functor, Ord, Show, Traversable)
+
+pvar :: a -> Pattern a
+pvar = PVal . PVar
+
+pcon :: Q Name :$ ValuePattern a -> Pattern a
+pcon = PVal . PCon
+
+
+fill :: Traversable t => (b -> (b, c)) -> b -> t a -> (b, t c)
+fill f = mapAccumL (const . f)
 
 
 -- Term values
